Move menu beneath main
