+++ src/main/menu/macos-menu.ts
